Try u128 to solve "big primes"

+++ exercism/rust/diffie-hellman/src/lib.rs
@@ -4,7 +4,7 @@ pub fn private_key(p: u64) -> u64 {
     rand::thread_rng().gen_range(2..p)
 }
 
-fn modular_exponentiation(x: u64, e: u64, m: u64) -> u64 {
+fn modular_exponentiation(x: u128, e: u128, m: u128) -> u128 {
     if e == 0 {
         1
     } else if e == 1 {
@@ -17,9 +17,9 @@ fn modular_exponentiation(x: u64, e: u64, m: u64) -> u64 {
 }
 
 pub fn public_key(p: u64, g: u64, a: u64) -> u64 {
-    modular_exponentiation(g, a, p)
+    modular_exponentiation(g as u128, a as u128, p as u128) as u64
 }
 
 pub fn secret(p: u64, b_pub: u64, a: u64) -> u64 {
-    modular_exponentiation(b_pub, a, p)
+    modular_exponentiation(b_pub as u128, a as u128, p as u128) as u64
 }
